Registration Finaly

+++ DildoShop/DildoShop.Server/Models/User.cs
@@ -6,6 +6,6 @@
         public string Login { get; set; }
         public string PasswordHash { get; set; }
         public string Gender { get; set; } 
-        public float PreferredDildoSize { get; set; } 
+        public int PreferredDildoSize { get; set; } 
     }
 }
