+++ DildoShop/DildoShop.Server/Models/UserRegistrationDto.cs
@@ -5,6 +5,6 @@
         public string Login { get; set; }
         public string Password { get; set; }
         public string Gender { get; set; }
-        public float PreferredDildoSize { get; set; }
+        public int PreferredDildoSize { get; set; }
     }
 }
